Update dependency versions only when the original version matches
This changes the behaviour so that when a matching GA is found the
version will not be updated unless the original version also matches.

This fixes erroneous updates which happens when a module of a project intentionally includes a dependency on an older version of another module of the same project
(for example, backwards compatibility tests).