[issue #446] Fix cases where we have an alphanumeric qualifier without delimiter
Previous behaviour was if the version qualifier contains an alpha part with a number at the end,
the parser interprets the number portion as a build number and reformats it with a dash
between the descriptor and the build number.  This change fixes cases where we don't
want to add the dash into the qualifier.