Allow `dependencyOverride`s to reference a remote BOM as a version
This allows for more than one `dependencyManagement` BOM to be used when they
would otherwise conflict. The `dependencyOverride`/`dependencyExclusion`s limit
these "extra BOMs" effect to their requested modules/artifacts only.

The core RH-SSO server needs to align to EAP 7.x, but it also must build an
adapter that compiles against EAP 6.x. Before this change, the way to do this
would be:

```
-DdependencyManagement=org.jboss.eap:jboss-eap-parent:7.2.0.GA-redhat
-DdependencyRelocations.org.wildfly:@org.jboss.eap:=7.2.0.GA-redhat
-DdependencyOverride.org.jboss.as:*@*=7.5.20.Final-redhat-1
-DdependencyOverride.org.jboss.web:jbossweb@*=7.5.28.Final-redhat-1
...
```

Which works, but requires manually synchronizing the dependencyOverrides. With
this change, this is how it'd look:

```
-DdependencyManagement=org.jboss.eap:jboss-eap-parent:7.2.0.GA-redhat
-DdependencyRelocations.org.wildfly:@org.jboss.eap:=7.2.0.GA-redhat
-DdependencyManagement.eap6=org.jboss.bom:eap6-supported-artifacts:6.4.21.GA
-DdependencyOverride.org.jboss.as:*@*=eap6
-DdependencyOverride.org.jboss.web:jbossweb@*=eap6
...
```

Which results in less churn and manual work.